tests/test_data/requirements_generate.txt
reportlab>=3.6.0
Pillow>=9.0.0
fpdf2>=2.5.0
